Fix item selection
Currently when you select anything, the body type switches back to "Male".
When you then select "Female" for instance, it is not processed. Only after a second click it does process.

I don't know why this happens exactly, but I tracked it down to this change that was recently made.
Reverting this fixes the issue.